Drop IntoCStr in favor of CStrBuf::from_vec, redesign errors
The trait was problematic in that it lost the original value
in case of failure. Now the failing conversion will return an
IntoCStrError which can be moved into the vector of original bytes.

It does not make sense to maintain a polymorphic trait as long
as the only type we can practically move from is a Vec, or anything
that cheaply translates into Vec. Another consuming conversion could
be provided for convenience, e.g. from_string(), but this would bloat
the interface, it can be trivially done by combining other functions,
and it's better to keep the design symmetric against the Vec falling
out of the failure path.

The CStrError got renamed to NulError and made into a simple struct,
because it is only ever going to indicate one error condition.