Recast the crate as c_str_macro 1.0.0
Most of the stuff defined in c_string was dodgy, obsolete and not used
much. c_str!, however, is quite popular. It's not nice to make people
remember to use #[no_link] if they don't have a use for the other cruft.
It would be wrong to release the macro-only crate under the old
name, too, as the name no longer matches the purpose and shedding most of
the functionality might still surprise some people who use "*" as
the dependency version.